Allow extra java options to be passed to `OPENSEARCH_JAVA_OPTS`
This allows any required options to be passed to the `OPENSEARCH_JAVA_OPTS`
flag. I have also renamed `jvm` variable to `javaopts` as this is where it is
actually passed to. Note `jvm` is still supported to be passed through from the nodepool spec.

I have not configured this to be supported at the bootstrap image, as ultimately I don't think
it will be useful here, I may be worng and am happy to be corrected.

Reason for allowing "arbitary" java options is because as features move from "experimental" to
"production" we do not want to have to keep changing the code to remove the experimental flags.

The flip side to this is if you want to try "experimental" feature you need to explicitly enable
them, in the `javaopts` flag.

Signed-off-by: Rhys Evans <rhys.rr@googlemail.com>