Update README.md
We got many requests during KubeCon from people interested in using the OMC along with the Operator, and mentioned it would have been helpful if it had been included in the same page as the Operator. Added here so users can easily learn how to install it and see the official page.