Fix #623. Use http port(default 9200) for curl and node communication port(9300) for securityadmin.sh for version 1.X.
For Opensearch version 2.X securityadmin.sh is using the http port (default 9200). However, for version 1.X securityadmin.sh is using the port 9300, which is normally used for node communication in a binary protocol. Because we use curl on the same port to check if the service is ready, before we start securityadmin.sh, curl was also using port 9300 and was never successful, because of the binary protocol. Therefore, the bootstrap pod run forever.

The solution was to split the ports for curl and securityadmin.sh, so curl uses the actual http port.

More details can be found here:
- https://forum.opensearch.org/t/securityadmin-sh-uses-http-port-9200-and-not-transport-port-9300-in-opensearch-2-0-0/9760

- https://discuss.elastic.co/t/what-are-ports-9200-and-9300-used-for/238578

Signed-off-by: Stefan Heilig <Stefan.Heilig@maibornwolff.de>