Introduce K8sClient abstraction for easier testing (#592)
This PR introduces a new interface `K8sClient` as an abstraction over
the existing `client.Client` and `ResourceReconciler`. The goals are to
reduce some boilerplate code and make it easier to unittest the code
with mocks instead of using the slow envtest approach everywhere.

Changes:
* New interface `K8sClient` with `K8sClientImpl` as default
implementation that wraps all kubernetes client access
* Introduce [mockery](https://vektra.github.io/mockery/) as helper
library to generate mocks
* Refactor all tests in `pkg/reconcilers` to use a `MockK8sClient`

With these changes it should be easier to write unittests for the
reconcilers in the future. Also the runtime of the tests should be
significantly reduced.

---------

Signed-off-by: Sebastian Woehrl <sebastian.woehrl@maibornwolff.de>