Add warning about cert expiration (#725)
### Description
This PR adds a warning to make users aware of the 1-year validity time
of operator-generated certificates. When documentation for renewing the
certificates is available, it would be good to add link to it from here.

### Issues Resolved
N/A

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

---------

Signed-off-by: Trond Hindenes <trond@hindenes.com>
Signed-off-by: Prudhvi Godithi <pgodithi@amazon.com>
Co-authored-by: Prudhvi Godithi <pgodithi@amazon.com>
Signed-off-by: Sunil <sunils@gmail.com>