Changed the structure of the cron condition for ISM policy CRD (#838)
### Description
The cron condition for ISM policy CRDs was not working since it had the
wrong structure. The structure in the operator was defined as

```
...
"conditions": {
  "cron": {
    "expression": "* 17 * * SAT",
    "timezone": "America/Los_Angeles"
  }
}
...
```

while the OpenSearch API expects

```
...
"conditions": {
  "cron": {
    "cron": {
      "expression": "* 17 * * SAT",
      "timezone": "America/Los_Angeles"
    }
  }
}
...
```

This PR should resolve the issue.

I also ran into a few odd details in the code where we checked if an
error was nil when it would always be nil so I have moved those lines
out of the if-statement. A few error messages were capital first letter
which was different from the rest of the codebase so I fixed those.
Lastly, my linter (default Go extension for VSCode) swapped some import
orders. Let me know if those changes are alright.

### Issues Resolved
I don't think this PR solves any current issues

### Check List
- [x] Commits are signed per the DCO using --signoff 
- [ ] Unittest added for the new/changed functionality and all unit
tests are successful
- [ ] Customer-visible features documented
- [x] No linter warnings (`make lint`)

If CRDs are changed:
- [x] CRD YAMLs updated (`make manifests`) and also copied into the helm
chart
- [x] Changes to CRDs documented - **(there was no documentation for
this part of the CRD, but I am happy to make one if you would like)**

Please refer to the [PR
guidelines](https://github.com/opensearch-project/opensearch-k8s-operator/blob/main/docs/developing.md#submitting-a-pr)
before submitting this pull request.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

Signed-off-by: rkthtrifork <rkth@trifork.com>