Enhanced the affinity for opensearch-cluster to have a default podAntiAffinity
This PR helps to add a default podAntiAffinity rule in each nodePool
component with the cluster-name as selector to schedule the replicas on
different nodes on Kubernetes.

All components using the same rule, thus each replica from nodePool
components schedules on different nodes. This can be improved on
component wise in nodePool in future if needed.

Fixes opensearch-project#285

- [x] Commits are signed per the DCO using --signoff
- [x] Unittest added for the new/changed functionality and all unit
tests are successful
- [x] Customer-visible features documented
- [x] No linter warnings (`helm lint` and `helm template`)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

Signed-off-by: Rajesh M <rajesh.m101090@gmail.com>