feat(pvc): Allow custom labels and annotations on PVCs
This commit introduces two new fields, `labels` and `annotations`, to the persistence configuration for PVCs.

This allows users to specify their own metadata on the PersistentVolumeClaims created by the operator. This is necessary for integrations with other tools, such as storage encryption systems or cost-monitoring platforms that rely on specific annotations or labels.

Example usage:

```yaml
spec:
  nodePools:
  - component: data
    persistence:
      pvc:
        labels:
          team: "backend-data"
        annotations:
          test.io/crypt-key-id: "your-kms-key-id"
```

Signed-off-by: Okkan SERIN <okkan.serin@unblu.com>