Fix: Don't force ssl.verificationMode to none (#1061)
### Description
Currently, `opensearch.ssl.verificationMode` is forcibly set to `none`
even if it's supplied via `additionalConfig`. This PR fixes this issue
by not overriding it when it's provided by the user.

### Issues Resolved
Closes #1060

### Check List
- [x] Commits are signed per the DCO using --signoff 
- [ ] Unittest added for the new/changed functionality and all unit
tests are successful
- [ ] Customer-visible features documented
- [x] No linter warnings (`make lint`)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: amrut-asm <amrutrn1@gmail.com>