Allow the opensearch operator to watch multiple namespaces
We keep the original `-watch-namespace` flag, to ensure backwards compatibility.
We simply split the value over any comma, and populate the cache for
each namespace in the csv.

Note: Because the `watchNamespace` variable was being tested for
emptiness _before_ `flag.Parse()` was being called, it was always empty,
causing the operator to _always_ watch all namespaces in the cluster.
This is no longer the case.

Fixes #374

Signed-off-by: Balthazar Rouberol <br@imap.cc>