Fix useRoleBindings to properly use Role instead of ClusterRole
Fixes #1130

When useRoleBindings is set to true, the operator should create Role
and RoleBinding resources instead of ClusterRole and ClusterRoleBinding
to allow installation with namespace-level privileges.

This fix updates the following templates to conditionally use Role when
useRoleBindings is true:

- opensearch-operator-manager-role-cr.yaml: Now creates Role when
  useRoleBindings is true, otherwise creates ClusterRole
- opensearch-operator-manager-rolebinding.yaml: Now references Role
  when useRoleBindings is true
- opensearch-operator-metrics-reader-cr.yaml: Now creates Role when
  useRoleBindings is true, otherwise creates ClusterRole
- opensearch-operator-proxy-role-cr.yaml: Now creates Role when
  useRoleBindings is true, otherwise creates ClusterRole
- opensearch-operator-proxy-rolebinding.yaml: Now references Role when
  useRoleBindings is true

This enables users with only namespace admin privileges to install the
operator successfully.