fix: resolve deadlocks when upgrading to 3.0.0 (#1217)
### Description
Implements various fixes to allow the upgrade from Opensearch 2.19.x to
3.x.
In the following paragraphs I will briefly explain the changes.
I will also make some minimal examples to explain the issues, referring
to a 3-node cluster where all nodes have data + cluster_manager roles.

#### Deadlock caused by UNASSIGNED shards due to `node_version` mismatch
Consider an index `foo` with 1 shard and 2 replicas.
After upgrading at least one node, the shard's primary is assigned to an
upgraded node.
Replicas cannot be assigned to nodes with an older version, therefore
they remain unassigned and the cluster is stuck in a YELLOW state. In
particular it gets stuck at `CheckClusterStatusForRestart`.

| index | shard | pri/rep | state | node | version |
|--------|--------|--------|--------|--------|--------|
| foo | 0 | p | STARTED | node-0 | 3.0.0 |
| foo | 0 | r | STARTED | node-1 | 3.0.0 |
| foo | 0 | r | UNASSIGNED |  | 2.19.4 |

**Solution**
If the cluster is YELLOW and ALL unassigned replicas are stuck in this
state, allow the upgrade to continue. Note that we ensure the UNASSIGNED
is a replica and not a primary, making sure that the shard has at least
a primary in a STARTED state, and it's on an already-upgraded node that
won't be restarted. Therefore we can safely restart the next pod.

#### Deadlock caused by ASSIGNED shards that cannot be moved due to
`node_version` mismatch (#1153)
Consider an index `foo` with 1 shard and 1 replicas.
After upgrading at least one node, the shard's primary is assigned to an
upgraded node.
If `drainNodes` is enabled, when the cluster tries to drain the next
node (node-1), it stalls because of a `node_version` mismatch, and the
node won't be able to fully drain.
Here the cluster is GREEN, but the operator gets stuck at
`PreparePodForDelete`.

| index | shard | pri/rep | state | node | version |
|--------|--------|--------|--------|--------|--------|
| foo | 0 | p | STARTED | node-0 | 3.3.0 |
| foo | 0 | r | STARTED | node-1 | 2.19.4 |
|  |  |  |  | node-2 | 2.19.4 |

**Solution**
If `drainNodes` is enabled, allow a node to be deleted anyway if ALL the
shards it contains are replicas stuck for this reason.
As before, the fact that these are replicas and that the primary is on
an upgraded node means that it's safe to restart the pod.

> NOTE: if this is merged, #1211 is not necessary, as they address the
same problem in different ways

#### Deadlock caused by `archived` settings (#1165)
This is related to
https://github.com/opensearch-project/OpenSearch/issues/18515 , which
despite being closed is still occurs as of 18/12/2025.

- Some settings set by default by the index state management plugin have
been removed in 3.0.0
- During the upgrade, the settings are archived, by prefixing them with
`archived`
- The already upgraded 3.x nodes tolerate `archived` settings
- The nodes <= 2.19.4 do not tolerate them on cluster settings updates,
throwing a 400 error
- Therefore when modifying allocation settings (to drain nodes and
enable/disable replica allocation), the operation will fail indefinetly,
causing a deadlock

The only way to get the process unstuck is to delete these settings.

**Solution**
The operator deletes the offending settings before restarting a Pod.
I don't _love_ this solution, as it's quite defensive, but without it
even the simple test "deploy 2.19.4 and update it to 3.3.0" wouldn't be
able to complete without manual intervention.

#### Functional tests
I have also applied the following fixes to make the `deploy-and-upgrade`
test work:
- lower the disk watermarks: this caused the cluster to get stuck and
was already implemented in the helm tests
- check that all replicas are both up-to-date and ready to pass the
test: before this change the test terminated as soon as the pod for the
last replica was created, without waiting for it to become ready

I have tested it locally, hope it also works on CI :wink: 

#### Logging improvements
Inside the upgrader, I have added some more logging in case of error,
which helped me while debugging.
Before, many errors were simply propagated upstream and never logged.

### Issues Resolved
Fixes #1153
Fixes #1165

### Check List
- [x] Commits are signed per the DCO using --signoff 
- [x] Unittest added for the new/changed functionality and all unit
tests are successful
- [ ] Customer-visible features documented
- [ ] No linter warnings (`make lint`)

If CRDs are changed:
- [ ] CRD YAMLs updated (`make manifests`) and also copied into the helm
chart
- [ ] Changes to CRDs documented

Please refer to the [PR
guidelines](https://github.com/opensearch-project/opensearch-k8s-operator/blob/main/docs/developing.md#submitting-a-pr)
before submitting this pull request.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

---------

Signed-off-by: Alessandro Fulgini <fuljo97@gmail.com>